Added a check for guest name missing.

+++ virsh-cdrom.sh
@@ -16,6 +16,14 @@ then
    echo "Unmount: sh virsh-cdrom.sh unmount GUEST_NAME"
    exit
 fi
+if [[ -z "$GUEST" ]]
+then
+   echo "Error: Guest name undefined"
+   echo "Example:"
+   echo "Mount: sh virsh-cdrom.sh mount GUEST_NAME ISO_NAME"
+   echo "Unmount: sh virsh-cdrom.sh unmount GUEST_NAME"
+   exit
+fi
 if [ "$ACTION" == mount ]; then
 	# shutdown the guest
 	/usr/bin/virsh destroy $GUEST
